// webclient/src/gen/message.ts
export enum Chroma {
  I420 = 0,
  I444 = 1,
}
// ...
export interface EncodedVideoFrame {
  data: Uint8Array;
  key: boolean;
  pts: bigint;
}

export interface EncodedVideoFrames {
  frames: EncodedVideoFrame[];
}

export interface RGB {
  compress: boolean;
}

/** planes data send directly in binary for better use arraybuffer on web */
export interface YUV {
  compress: boolean;
  stride: number;
}

export interface VideoFrame {
  union?:
    | { $case: "vp9s"; vp9s: EncodedVideoFrames }
    | { $case: "rgb"; rgb: RGB }
    | { $case: "yuv"; yuv: YUV }
    | { $case: "h264s"; h264s: EncodedVideoFrames }
    | { $case: "h265s"; h265s: EncodedVideoFrames }
    | { $case: "vp8s"; vp8s: EncodedVideoFrames }
    | { $case: "av1s"; av1s: EncodedVideoFrames }
    | undefined;
  display: number;
}
// ...
export interface CodecAbility {
  vp8: boolean;
  vp9: boolean;
  av1: boolean;
  h264: boolean;
  h265: boolean;
}
// ...
export interface SupportedDecoding {
  ability_vp9: number;
  ability_h264: number;
  ability_h265: number;
  prefer: SupportedDecoding_PreferCodec;
  ability_vp8: number;
  ability_av1: number;
  i444?: CodecAbility | undefined;
  prefer_chroma: Chroma;
}

export enum SupportedDecoding_PreferCodec {
  Auto = 0,
  VP9 = 1,
  H264 = 2,
  H265 = 3,
  VP8 = 4,
  AV1 = 5,
}
// ...
export const EncodedVideoFrame: MessageFns<EncodedVideoFrame> = {
  encode(message: EncodedVideoFrame, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.data.length !== 0) {
      writer.uint32(10).bytes(message.data);
    }
    if (message.key !== false) {
      writer.uint32(16).bool(message.key);
    }
    if (message.pts !== 0n) {
      if (BigInt.asIntN(64, message.pts) !== message.pts) {
        throw new globalThis.Error("value provided for field message.pts of type int64 too large");
      }
      writer.uint32(24).int64(message.pts);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EncodedVideoFrame {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEncodedVideoFrame();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.data = reader.bytes();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.key = reader.bool();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.pts = reader.int64() as bigint;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<EncodedVideoFrame>, I>>(base?: I): EncodedVideoFrame {
    return EncodedVideoFrame.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EncodedVideoFrame>, I>>(object: I): EncodedVideoFrame {
    const message = createBaseEncodedVideoFrame();
    message.data = object.data ?? new Uint8Array(0);
    message.key = object.key ?? false;
    message.pts = (object.pts !== undefined && object.pts !== null) ? BigInt(object.pts) : 0n;
    return message;
  },
};

function createBaseEncodedVideoFrames(): EncodedVideoFrames {
  return { frames: [] };
}

export const EncodedVideoFrames: MessageFns<EncodedVideoFrames> = {
  encode(message: EncodedVideoFrames, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    for (const v of message.frames) {
      EncodedVideoFrame.encode(v!, writer.uint32(10).fork()).join();
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): EncodedVideoFrames {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseEncodedVideoFrames();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 10) {
            break;
          }

          message.frames.push(EncodedVideoFrame.decode(reader, reader.uint32()));
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<EncodedVideoFrames>, I>>(base?: I): EncodedVideoFrames {
    return EncodedVideoFrames.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<EncodedVideoFrames>, I>>(object: I): EncodedVideoFrames {
    const message = createBaseEncodedVideoFrames();
    message.frames = object.frames?.map((e) => EncodedVideoFrame.fromPartial(e)) || [];
    return message;
  },
};
// ...
export const CodecAbility: MessageFns<CodecAbility> = {
  encode(message: CodecAbility, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.vp8 !== false) {
      writer.uint32(8).bool(message.vp8);
    }
    if (message.vp9 !== false) {
      writer.uint32(16).bool(message.vp9);
    }
    if (message.av1 !== false) {
      writer.uint32(24).bool(message.av1);
    }
    if (message.h264 !== false) {
      writer.uint32(32).bool(message.h264);
    }
    if (message.h265 !== false) {
      writer.uint32(40).bool(message.h265);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): CodecAbility {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseCodecAbility();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.vp8 = reader.bool();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.vp9 = reader.bool();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.av1 = reader.bool();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.h264 = reader.bool();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.h265 = reader.bool();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<CodecAbility>, I>>(base?: I): CodecAbility {
    return CodecAbility.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<CodecAbility>, I>>(object: I): CodecAbility {
    const message = createBaseCodecAbility();
    message.vp8 = object.vp8 ?? false;
    message.vp9 = object.vp9 ?? false;
    message.av1 = object.av1 ?? false;
    message.h264 = object.h264 ?? false;
    message.h265 = object.h265 ?? false;
    return message;
  },
};
// ...
export const SupportedDecoding: MessageFns<SupportedDecoding> = {
  encode(message: SupportedDecoding, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.ability_vp9 !== 0) {
      writer.uint32(8).int32(message.ability_vp9);
    }
    if (message.ability_h264 !== 0) {
      writer.uint32(16).int32(message.ability_h264);
    }
    if (message.ability_h265 !== 0) {
      writer.uint32(24).int32(message.ability_h265);
    }
    if (message.prefer !== 0) {
      writer.uint32(32).int32(message.prefer);
    }
    if (message.ability_vp8 !== 0) {
      writer.uint32(40).int32(message.ability_vp8);
    }
    if (message.ability_av1 !== 0) {
      writer.uint32(48).int32(message.ability_av1);
    }
    if (message.i444 !== undefined) {
      CodecAbility.encode(message.i444, writer.uint32(58).fork()).join();
    }
    if (message.prefer_chroma !== 0) {
      writer.uint32(64).int32(message.prefer_chroma);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): SupportedDecoding {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseSupportedDecoding();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.ability_vp9 = reader.int32();
          continue;
        }
        case 2: {
          if (tag !== 16) {
            break;
          }

          message.ability_h264 = reader.int32();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.ability_h265 = reader.int32();
          continue;
        }
        case 4: {
          if (tag !== 32) {
            break;
          }

          message.prefer = reader.int32() as any;
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.ability_vp8 = reader.int32();
          continue;
        }
        case 6: {
          if (tag !== 48) {
            break;
          }

          message.ability_av1 = reader.int32();
          continue;
        }
        case 7: {
          if (tag !== 58) {
            break;
          }

          message.i444 = CodecAbility.decode(reader, reader.uint32());
          continue;
        }
        case 8: {
          if (tag !== 64) {
            break;
          }

          message.prefer_chroma = reader.int32() as any;
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<SupportedDecoding>, I>>(base?: I): SupportedDecoding {
    return SupportedDecoding.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<SupportedDecoding>, I>>(object: I): SupportedDecoding {
    const message = createBaseSupportedDecoding();
    message.ability_vp9 = object.ability_vp9 ?? 0;
    message.ability_h264 = object.ability_h264 ?? 0;
    message.ability_h265 = object.ability_h265 ?? 0;
    message.prefer = object.prefer ?? 0;
    message.ability_vp8 = object.ability_vp8 ?? 0;
    message.ability_av1 = object.ability_av1 ?? 0;
    message.i444 = (object.i444 !== undefined && object.i444 !== null)
      ? CodecAbility.fromPartial(object.i444)
      : undefined;
    message.prefer_chroma = object.prefer_chroma ?? 0;
    return message;
  },
};
// ...
type Builtin = Date | Function | Uint8Array | string | number | boolean | bigint | undefined;

export type DeepPartial<T> = T extends bigint ? string | number | bigint
  : T extends Builtin ? T : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends { $case: string } ? { [K in keyof Omit<T, "$case">]?: DeepPartial<T[K]> } & { $case: T["$case"] }
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}